added comments to virtual memory management and separated vma into separate file
