+++ os/kernel/src/memory/mod.rs
@@ -2,6 +2,8 @@ pub mod alloc;
 pub mod physical;
 pub mod r#virtual;
 pub mod nvmem;
+pub mod vma;
+
 
 #[derive(Clone, Copy)]
 pub enum MemorySpace {
